+++ src/prefect/cli/agent.py
@@ -116,6 +116,9 @@ def agent():
     help="Host paths for Docker bind mount volumes attached to each Flow runtime container.",
     hidden=True,
 )
+@click.option(
+    "--network", help="Add containers to an existing docker network", hidden=True,
+)
 @click.pass_context
 def start(
     ctx,
@@ -132,6 +135,7 @@ def start(
     import_path,
     show_flow_logs,
     volume,
+    network,
     max_polls,
 ):
     """
@@ -172,6 +176,7 @@ def start(
                                 Defaults to pulling if not provided
         --volume        TEXT    Host paths for Docker bind mount volumes attached to each Flow runtime container.
                                 Multiple values supported e.g. `--volume /some/path --volume /some/other/path`
+        --network       TEXT    Add containers to an existing docker network
 
     \b
     Kubernetes Agent Options:
@@ -228,6 +233,7 @@ def start(
                 no_pull=no_pull,
                 show_flow_logs=show_flow_logs,
                 volumes=list(volume),
+                network=network,
             ).start()
         elif agent_option == "fargate":
             from_qualified_name(retrieved_agent)(